requirements.txt
chroma-hnswlib
numpy
pandas
jupyter
seaborn
matplotlib
scipy
statsmodels
scikit-learn
python-dotenv
langchain
openai
huggingface_hub
langchain_experimental
chromadb
tiktoken
streamlit